src/sprint2_postgres/requirements.txt
psycopg2-binary==2.9.6
PyMySQL==1.0.2
python-dotenv==0.15.0
jmespath==1.0.1